requirements.txt
py-cord
deepgram-sdk
pynacl
python-dotenv
# whisper-cpp-python
pywhispercpp
pydub